refactor(guestbook): colocate modal forms
