+++ src/app/(preview)/_components/index.tsx
@@ -1,3 +1,1 @@
-export { CreateGuestbookForm } from "./CreateGuestbookForm";
-export { DeleteGuestbookForm } from "./DeleteGuestbookForm";
 export { GuestbookModal } from "./GuestbookModal";
